Add 0.8 migration (#278)
* feat: upgrade contracts to solidity 0.8

* feat: continue with 0.8 migration

- Remove audits
- Remove NFTDescriptor stuff because couldn't easily work through
  a weird assembly stacktoodeep issue. I think it has to do with
  ABIcoderv2

* fix: package.json

* fix: tests for 0.8 migration

* fix: clean up comment

* feat: upgrade to 0.8.15

* fix: bump version

* Fix code style issues with Prettier

* fix: improve gas for liquidityamounts

* Fix code style issues with Prettier

* fix: review comments

- add back nft descriptor tests
- remove xits
- improve ticklens gas

* Fix code style issues with Prettier

+++ contracts/libraries/PoolTicksCounter.sol
@@ -23,10 +23,10 @@ library PoolTicksCounter {
         {
             // Get the key and offset in the tick bitmap of the active tick before and after the swap.
             int16 wordPos = int16((tickBefore / self.tickSpacing()) >> 8);
-            uint8 bitPos = uint8((tickBefore / self.tickSpacing()) % 256);
+            uint8 bitPos = uint8(int8((tickBefore / self.tickSpacing()) % 256));
 
             int16 wordPosAfter = int16((tickAfter / self.tickSpacing()) >> 8);
-            uint8 bitPosAfter = uint8((tickAfter / self.tickSpacing()) % 256);
+            uint8 bitPosAfter = uint8(int8((tickAfter / self.tickSpacing()) % 256));
 
             // In the case where tickAfter is initialized, we only want to count it if we are swapping downwards.
             // If the initializable tick after the swap is initialized, our original tickAfter is a
